requirements.txt
streamlit
google-generativeai
reportlab
markdown
html.parser